requirements.txt
mlx-whisper>=0.4.0
sounddevice>=0.4.6
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
PyQt6>=6.5.0
